Bonjour - Fix package, update script optimizations, install improvements (#3891)
* Pin download URL to specific version of iTunes

Use of a canonical download URL makes the remote checksum subject to
change, which will break checksum verification and require package
updates. Using a specific release should provide better stability.

* Optimize Checksum32 calculation to avoid a second 32-bit installer download

* Ensure Checksum64's algorithm matches that of ChecksumType64

* Replace Set-Location calls with Push/Pop-Location

This should make it easier to return to the starting directory, should
the script error out or stall prior to completing.

* Remove wildcard pattern from target filename pattern

A wildcard is not be necessary since we only check the 32-bit installer.

* Extract only the necessary files from installer

This should avoid extracting unnecessary files and cluttering up TEMP

* Use 7z switch for "assume Yes on all queries"

This should avoid a run-time prompt for overwriting the extracted files,
which stalls the update script if TEMP is not cleaned between runs.

* Increment package version

* Add shim ignore file for SetupAdmin.exe

This binary is packaged within the iTunes MSI and is not needed by users.

* Use Install-ChocolateyInstallPackage for Bonjour MSI execution

Install-ChocolateyPackage is not intended for local file execution, and
will print about an "attempt to use original download file name" failing.